feat: 장바구니 가격 업데이트 로직 분리 및 통합
장바구니 내 상품 가격 업데이트 관련 로직을 `cartPriceHandlers.ts`로 분리하여 코드의 가독성과 유지보수성을 향상시켰습니다. `main.basic.ts`에서 가격 업데이트 관련 주석을 수정하고, 상품 타이머 이벤트 핸들러에서 `doUpdatePricesInCart` 함수를 호출하도록 변경하여 장바구니 가격 업데이트 기능을 명확하게 관리할 수 있게 되었습니다.
